Выполнит рефакт-г компонента сортировки
